requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
esda==2.8.1
fonttools==4.61.1
geopandas==1.1.2
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
libpysal==4.14.1
MarkupSafe==3.0.3
matplotlib==3.10.8
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.0
pillow==12.1.0
platformdirs==4.5.1
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
python-dateutil==2.9.0.post0
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
shapely==2.1.2
SimpleITK-SimpleElastix==2.5.0.dev49
six==1.17.0
soupsieve==2.8.3
specio-py310==0.1.0.post2
threadpoolctl==3.6.0
tifffile==2026.1.28
typing_extensions==4.15.0
urllib3==2.6.3
